Show notification messages when calculated properties change.
We want the player to know when they cross certain thresholds or
important calculated values. The main one that triggered this is wanting
to show an error message when the player is using a weapon whose heft
is too high.

But showing that message at all the right times is difficult. The heft
scale is based on the current weapon, hero strength, and weight of all
equipment. Strength is in turn based on level. When any of those change,
the heft scale might tip above or below 1.0 and we need to show the
message.

I introduced a Property class to let us track that. Basically, it stores
a cache of the previously calculated value. So we can refresh it to its
current value and then determine if the new value is actually different
and show a message then.

Then I went through and made level, stats, and heft scale all use this
class. There is a single function that refreshes all of them and shows
any notifications needed.

All of the actions that change the hero's equipment call this. We also
call it when experience changes. It's not very fine-grained -- it
refreshes every single property regardless of what changed, but that's
probably fine.