Revamp the item screen.
Instead of a two-column "from" and "to" view, only shows a single list
of items. This gives more room for the list and also allows the
inspector to be shown.

This changes the interaction entirely, so I basically rewrote the whole
thing. There are a couple of relatively minor TODOs, but most things
should be working.