Optimize passage carving.
Whenever a tile is filled, we need to recalculate the BFS map to see if
any open tiles are now unreachable. Previously, it recalculated the
whole stage from scratch each time. That's obviously slow since most
tiles are unaffected.

This adds an incrementally-updateable BFS map so that when a tile is
filled, only other affected tiles are recalculated. Makes carving
passages more than twice as fast.