Fix how ResourceSet handles resources in non-leaf tags.
When I made decor use ResourceSet, I had the wrong mental model for what
it means for a resource to be directly in some non-leaf tag. For decor,
the tags are like:

  built/room/dungeon
  built/room/keep
  catacomb
  cave/glowing-moss
  water

Most of the room decor is tagged "room". When decorating a dungeon, we
look for "dungeon" decor. The model for these is that anything in a
parent tag is effectively in all of the children. So anything tagged
"room" can show up in both dungeons and keeps. I "fixed" ResourceSet to
implement that.

This is *not* the model that other uses of ResourceSet. For monsters,
you have:

  natural/bug/worm

A query for "worm" should not usually return cockroaches, which are
tagged "bug". They aren't worms.

The change to ResourceSet for decor was wrong. Instead, if a given piece
of decor should be equally common in two leaf tags, it should have both
of those tags.